fixed usage of UINT when not needed

+++ quicksort.c
@@ -13,9 +13,9 @@
 
 #define SWAP(x, y) do { typeof(x) SWAP = x; x = y; y = SWAP; } while (0)
 
-UINT partition(UINT* A, UINT lo, UINT hi){
-    UINT pivot = A[hi];
-    UINT i = lo;
+int partition(UINT* A, int lo, int hi){
+    int pivot = A[hi];
+    int i = lo;
     for (int j = lo; j < hi; ++j) {
         if (A[j] < pivot) {
             SWAP(A[i], A[j]);
@@ -28,7 +28,7 @@ UINT partition(UINT* A, UINT lo, UINT hi){
 
 int quicksort(UINT* A, int lo, int hi) {
     if (lo < hi){
-        UINT p = partition(A, (UINT) lo, (UINT) hi);
+        int p = partition(A, lo, hi);
         quicksort(A, lo, p - 1);
         quicksort(A, p + 1, hi);
     }
